a new mathjax API
