+++ packages/mathjax/src/node.js
@@ -7,13 +7,18 @@ const types = {
   tex: "tex"
 };
 
+import MathJaxContext, {
+  type MathJaxObject,
+  type MathJaxContextValue
+} from "./context";
+
 type Props = {
   inline: boolean,
   children: string,
   onRender: ?Function
 };
 
-class Node extends React.Component<Props, *> {
+class MathJaxNode_ extends React.Component<Props & MathJaxContextValue, null> {
   script: ?HTMLScriptElement;
   nodeRef: React.ElementRef<*>;
 
@@ -22,7 +27,7 @@ class Node extends React.Component<Props, *> {
     onRender: null
   };
 
-  constructor(props: Props) {
+  constructor(props: Props & MathJaxContextValue) {
     super(props);
 
     this.nodeRef = React.createRef();
@@ -40,21 +45,11 @@ class Node extends React.Component<Props, *> {
   /**
    * Update the jax, force update if the display mode changed
    */
-  componentDidUpdate(prevProps: Props) {
+  componentDidUpdate(prevProps: Props & MathJaxContextValue) {
     const forceUpdate =
       prevProps.inline !== this.props.inline ||
       prevProps.children !== this.props.children;
     this.typeset(forceUpdate);
-  }
-
-  /**
-   * Prevent update when the source has not changed
-   */
-  shouldComponentUpdate(nextProps: Props, nextState: *, nextContext: *) {
-    return (
-      nextProps.children !== this.props.children ||
-      nextProps.inline !== this.props.inline
-    );
   }
 
   /**
@@ -68,7 +63,11 @@ class Node extends React.Component<Props, *> {
    * Clear the jax
    */
   clear() {
-    const MathJax = this.context.MathJax;
+    const MathJax = this.props.MathJax;
+
+    if (!MathJax) {
+      return;
+    }
 
     if (!this.script) {
       return;
@@ -86,9 +85,9 @@ class Node extends React.Component<Props, *> {
    * @param { Boolean } forceUpdate
    */
   typeset(forceUpdate: boolean = false) {
-    const { MathJax } = this.context;
+    const { MathJax } = this.props;
 
-    if (!MathJax) {
+    if (!MathJax || !MathJax.Hub) {
       throw Error(
         "Could not find MathJax while attempting typeset! It's likely the MathJax script hasn't been loaded or MathJax.Context is not in the hierarchy"
       );
@@ -103,6 +102,10 @@ class Node extends React.Component<Props, *> {
     if (forceUpdate || !this.script) {
       this.setScriptText(text);
     }
+    // As an invariant of above, this shouldn't occur
+    if (!this.script) {
+      return;
+    }
 
     MathJax.Hub.Queue(MathJax.Hub.Reprocess(this.script, this.props.onRender));
   }
@@ -113,7 +116,7 @@ class Node extends React.Component<Props, *> {
    */
   setScriptText(text: *) {
     const inline = this.props.inline;
-    const type = types[this.context.input];
+    const type = types[this.props.input];
     if (!this.script) {
       this.script = document.createElement("script");
       this.script.type = `math/${type}; ${inline ? "" : "mode=display"}`;
@@ -139,8 +142,22 @@ class Node extends React.Component<Props, *> {
   }
 }
 
-Node.contextTypes = {
-  MathJax: PropTypes.object,
-  input: PropTypes.string
-};
-export default Node;
+class MathJaxNode extends React.PureComponent<Props, null> {
+  render() {
+    return (
+      <MathJaxContext.Consumer>
+        {({ MathJax, input }: MathJaxContextValue) => {
+          if (!MathJax) {
+            return null;
+          }
+
+          return (
+            <MathJaxNode_ {...this.props} input={input} MathJax={MathJax} />
+          );
+        }}
+      </MathJaxContext.Consumer>
+    );
+  }
+}
+
+export default MathJaxNode;
